Moved HashIndex, refactored the hash type computation module into the HashIndex code
