+++ pipeline.py
@@ -5,15 +5,13 @@ import traceback
 import time
 
 from pathlib import Path
-
 from pprint import pprint
 
+from log_config import setup_logging
 
 from sto_sister.pipeline import build_default_pipeline, PipelineContext
 from sto_sister.exceptions import SISTERError, PipelineError, StageError
-from sto_sister.hashindex import HashIndex
-from log_config import setup_logging
-
+from sto_sister.utils.hashindex import HashIndex
 from sto_sister.utils.image import load_image
 
 import traceback
